+++ src/NetCore/Web/Api/Program.cs
@@ -10,6 +10,7 @@ using E.Standard.Custom.Core.Abstractions;
 using E.Standard.Localization.Abstractions;
 using E.Standard.Security.App.Json;
 using E.Standard.Security.Cryptography.Abstractions;
+using E.Standard.WebApp.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -19,6 +20,7 @@ using System.Collections.Generic;
 
 var builder = WebApplication
                     .CreateBuilder(args)
+                    .SetAppLocalization(false)
                     .PerformWebgisApiSetup(args)
                     .AddWebgisApiConfiguration();