+++ src/NetCore/Web/Cms/Startup.cs
@@ -24,6 +24,8 @@ using E.Standard.Security.Cryptography.Extensions.DependencyInjection;
 using E.Standard.Security.Cryptography.Services;
 using E.Standard.Web.Extensions.DependencyInjection;
 using E.Standard.Web.Services;
+using E.Standard.WebApp.Abstraction;
+using E.Standard.WebApp.Extensions;
 using E.Standard.WebGIS.SubscriberDatabase.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Builder;
@@ -368,6 +370,8 @@ public class Startup
             }
 
             app.UseDatalinqCodeAuthentication();
+
+            app.RegisterApiEndpoints(typeof(Startup)).RegisterApiEndpoints(typeof(IApiEndpoint));
 
             app.UseMvc(routes =>
             {